Feature: Arguments, GUI, Priority
Added the features:

Arguments: This feature allows  the user to launch the program without GUI or with it.

GUI: Added a GUI in the terminal. This shows the current playing song and album. and allows the user to quickly open the logs and clear them. By default its On. but you can still launch the legacy no GUI version using the -s argument

Priority: Allows you to change what to prioritize. This can be Web or Plexamp.